# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MissionOut API",
    "description": "MissionOut API contract for dispatcher, team admin, and responder clients. The exported OpenAPI document in contracts/openapi.json is the cross-stack source of truth.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/teams/{team_public_id}/members": {
      "get": {
        "tags": [
          "teams"
        ],
        "summary": "List Team Members",
        "operationId": "list_team_members_teams__team_public_id__members_get",
        "parameters": [
          {
            "name": "team_public_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Team Public Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamMemberRead"
                  },
                  "title": "Response List Team Members Teams  Team Public Id  Members Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "teams"
        ],
        "summary": "Create Team Member",
        "operationId": "create_team_member_teams__team_public_id__members_post",
        "parameters": [
          {
            "name": "team_public_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Team Public Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamMemberCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMemberRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/teams/{team_public_id}/members/{membership_public_id}": {
      "patch": {
        "tags": [
          "teams"
        ],
        "summary": "Update Team Member",
        "operationId": "update_team_member_teams__team_public_id__members__membership_public_id__patch",
        "parameters": [
          {
            "name": "team_public_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Team Public Id"
            }
          },
          {
            "name": "membership_public_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Membership Public Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamMemberUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMemberRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "teams"
        ],
        "summary": "Delete Team Member",
        "operationId": "delete_team_member_teams__team_public_id__members__membership_public_id__delete",
        "parameters": [
          {
            "name": "team_public_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Team Public Id"
            }
          },
          {
            "name": "membership_public_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Membership Public Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/teams/{team_public_id}/devices": {
      "get": {
        "tags": [
          "teams"
        ],
        "summary": "List Team Devices",
        "operationId": "list_team_devices_teams__team_public_id__devices_get",
        "parameters": [
          {
            "name": "team_public_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Team Public Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceRead"
                  },
                  "title": "Response List Team Devices Teams  Team Public Id  Devices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceRead": {
        "properties": {
          "public_id": {
            "type": "string",
            "title": "Public Id"
          },
          "user_public_id": {
            "type": "string",
            "title": "User Public Id"
          },
          "user_name": {
            "type": "string",
            "title": "User Name"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "push_token": {
            "type": "string",
            "title": "Push Token"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "is_verified": {
            "type": "boolean",
            "title": "Is Verified"
          }
        },
        "type": "object",
        "required": [
          "public_id",
          "user_public_id",
          "user_name",
          "platform",
          "push_token",
          "last_seen",
          "is_active",
          "is_verified"
        ],
        "title": "DeviceRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TeamMemberCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "phone": {
            "type": "string",
            "title": "Phone",
            "default": ""
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "email",
          "roles"
        ],
        "title": "TeamMemberCreate"
      },
      "TeamMemberRead": {
        "properties": {
          "public_id": {
            "type": "string",
            "title": "Public Id"
          },
          "user_public_id": {
            "type": "string",
            "title": "User Public Id"
          },
          "team_public_id": {
            "type": "string",
            "title": "Team Public Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "granted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Granted At"
          }
        },
        "type": "object",
        "required": [
          "public_id",
          "user_public_id",
          "team_public_id",
          "name",
          "email",
          "phone",
          "roles",
          "is_active",
          "granted_at"
        ],
        "title": "TeamMemberRead"
      },
      "TeamMemberUpdate": {
        "properties": {
          "roles": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roles"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "TeamMemberUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}